+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.0.11
 
+- Support element custom event listener: `on_remove`
+- Support `observe` keyword to observe Model attributes in Views with a convenient DSL keyword.
+- Clean observers declared with `observe` keyword inside a component when `component.markup_root.remove` is called (or equivalent `component.remove` is called).
 - Update `glimmer_component` to render its content immediately without waiting for `Document.ready?` as that is not necessary given its container is added to the DOM immediately
 - Upgrade jquery-opal to 0.5.0 (latest version)
 - Button Counter Sample: `require 'glimmer-dsl-web/samples/regular/button_counter'`